Moved less work out of the controller.
Less work is now in the controller and now in the appropriate classes.
Validation is now done within the person class, and the family class
now takes care of calculating family taxes.

Furthermore, the tax class is now abstracted, and there's a new method
for calculating federal taxes. There's also a new method to both
calculate and set the afterTax var in one method, allowing for less
clunky implementation.